+++ apps/ledger-live-desktop/src/renderer/modals/UpdateFirmwareModal/steps/02-step-flash-mcu.tsx
@@ -9,7 +9,7 @@ import { isDeviceLocalizationSupported } from "@ledgerhq/live-common/manager/loc
 import firmwareUpdateMain from "@ledgerhq/live-common/hw/firmwareUpdate-main";
 import { withDevicePolling } from "@ledgerhq/live-common/hw/deviceAccess";
 import getDeviceInfo from "@ledgerhq/live-common/hw/getDeviceInfo";
-import { getEnv } from "@ledgerhq/live-common/env";
+import { getEnv } from "@ledgerhq/live-env";
 import TrackPage from "~/renderer/analytics/TrackPage";
 import Box from "~/renderer/components/Box";
 import FlashMCU from "~/renderer/components/FlashMCU";