+++ apps/ledger-live-mobile/src/screens/SyncOnboarding/UnlockDeviceDrawer.tsx
@@ -9,19 +9,25 @@ import QueuedDrawer from "../../components/QueuedDrawer";
 export type Props = {
   isOpen: boolean;
   onPress?: () => void;
+  /**
+   * Callback triggered when the user closes the drawer (by clicking on the backdrop or the close button)
+   * AND when the drawer is hidden.
+   * This is currently due to a legacy behavior of the BaseModal component. It might change in the future.
+   * You cannot know with just this callback if it's the user who closed the drawer.
+   */
   onClose?: () => void;
   device: Device;
 };
 
-const UnlockDeviceDrawer = ({ isOpen, device, onClose }: Props) => {
+const UnlockDeviceDrawer: React.FC<Props> = ({ isOpen, device, onClose }) => {
   const { t } = useTranslation();
   const { colors } = useTheme();
   const theme = colors.type as "dark" | "light";
 
   return (
     <QueuedDrawer isRequestingToBeOpened={isOpen} onClose={onClose} preventBackdropClick>
-      <Flex mb={250} pt={120}>
-        <ConnectYourDevice t={t} device={device} unresponsive={true} theme={theme} />
+      <Flex flexDirection="row">
+        <ConnectYourDevice t={t} device={device} isLocked theme={theme} />
       </Flex>
     </QueuedDrawer>
   );