+++ libs/ledger-live-common/src/apps/logic.test.ts
@@ -12,7 +12,7 @@ import {
 import { /*runAll,*/ runOneAppOp } from "./runner";
 import { deviceInfo155, mockListAppsResult, mockExecWithInstalledContext } from "./mock";
 // import { prettyActionPlan, prettyInstalled } from "./formatting";
-import { setEnv } from "../env";
+import { setEnv } from "@ledgerhq/live-env";
 import { Action } from "./types";
 
 setEnv("MANAGER_INSTALL_DELAY", 0);