+++ libs/ledger-live-common/src/families/tezos/synchronisation.ts
@@ -12,7 +12,7 @@ import { areAllOperationsLoaded, decodeAccountId } from "../../account";
 import type { Operation, Account } from "@ledgerhq/types-live";
 import api from "./api/tzkt";
 import type { APIOperation } from "./api/tzkt";
-import { getEnv } from "../../env";
+import { getEnv } from "@ledgerhq/live-env";
 
 function reconciliatePublicKey(
   publicKey: string | undefined,